+++ src/main/java/use_case/message_board/message/MessageInputData.java
@@ -2,5 +2,5 @@ package use_case.message_board.message;
 
 import entities.user.User;
 
-public record MessageInputData(int projectID, int messageID, User user, String newComment) {
+public record MessageInputData(long projectID, long messageID, User user, String newComment) {
 }
