+++ src/main/java/view/JLabelWithFont.java
@@ -4,7 +4,7 @@ import javax.swing.*;
 import java.awt.*;
 
 public class JLabelWithFont extends JLabel {
-    private final int defaultSize = 18;
+    private final long defaultSize = 18;
     public JLabelWithFont() {
         super();
         setFont(new Font("Times New Roman", Font.PLAIN, defaultSize));
@@ -15,17 +15,17 @@ public class JLabelWithFont extends JLabel {
         setFont(new Font("Times New Roman", Font.PLAIN, defaultSize));
     }
 
-    public JLabelWithFont(String text, int size) {
+    public JLabelWithFont(String text, long size) {
         super(text);
         setFont(new Font("Times New Roman", Font.PLAIN, size));
     }
 
-    public JLabelWithFont(String text, int style, int size) {
+    public JLabelWithFont(String text, long style, long size) {
         super(text);
         setFont(new Font("Times New Roman", style, size));
     }
 
-    public JLabelWithFont(String text, String name, int style, int size) {
+    public JLabelWithFont(String text, String name, long style, long size) {
         super(text);
         setFont(new Font(name, style, size));
     }
